pyproject.toml
[project]
name = "nunuk-agent"
version = "0.1.0"
description = "Build-your-own coding agent (course implementation)"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.15.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.0.1",
    "rich>=15.0.0",
]

[tool.uv]
package = false
